⚡️ Move analysis export zip generation to web worker
Closes #419

🚨 Address lints

🐛 Try to make idb.svelte.js importable in web worker

Revert "🐛 Try to make idb.svelte.js importable in web worker"

This reverts commit 6e7b3cd90596189f38b7c8f33a756753bea5b57d.

🐛 Fix retrieval of cropbox in implementation of analysis exporting

♻ Remove (almost) useless origin argument for generateResultsZip

🐛 EXIF writing: write bytestring using 32k-byte chunks instead of 65k

🐛 Use an OffscreenCanvas to crop images

🐛 Fix missing manuallyModified value for crop metadata

🐛 Fix observationMetadata and addValueLabels

🚨 Unexport METADATA_ZERO_VALUE

🐛 Add back progress bar

Fixes #420
