+++ src/lib/KeyboardShortcuts.svelte
@@ -12,7 +12,7 @@
 	 * @type {object}
 	 * @property {import('$lib/state.svelte.js').Keymap} binds
 	 * @property {boolean} [preventDefault=false] call e.preventDefault() before calling the handlers
-	 * @property {() => void} [openHelp] open modal listing all keybindings
+	 * @property {(() => void) | undefined} [openHelp] open modal listing all keybindings
 	 */
 
 	/** @type {Props} */
