+++ src/lib/i18n.js
@@ -59,8 +59,7 @@ export function percent(value, decimals = 0, { pad = 'none' } = {}) {
  * @returns {Promise<Record<string, number>>} map of language codes to number in [0, 1] indicating the completion percentage
  */
 export async function languagesCompletions() {
-	const { browser } = await import('$app/environment');
-	if (!browser) return {};
+	if (!localStorage) return {};
 
 	const cached = localStorage.getItem('languagesCompletions');
 	if (cached) {
