+++ src/lib/results.svelte.js
@@ -1,334 +1,12 @@
-/**
- * @import * as DB from './database';
- */
 import { ArkErrors } from 'arktype';
-import { strFromU8, strToU8, unzipSync, zip } from 'fflate';
-import { coordsScaler, toTopLeftCoords } from './BoundingBoxes.svelte';
-import { Schemas } from './database';
-import { downloadAsFile, stringifyWithToplevelOrdering } from './download';
-import { addExifMetadata } from './exif';
+import { strFromU8, unzipSync } from 'fflate';
 import * as db from './idb.svelte';
 import { imageIdToFileId, resizeToMaxSize, storeImageBytes } from './images';
-import {
-	addValueLabels,
-	METADATA_ZERO_VALUE,
-	metadataPrettyKey,
-	metadataPrettyValue,
-	observationMetadata,
-	serializeMetadataValue
-} from './metadata';
-import { isNamespacedToProtocol, removeNamespaceFromMetadataId } from './schemas/metadata';
+import { serializeMetadataValue } from './metadata';
 import { Analysis } from './schemas/results';
 import { uiState } from './state.svelte';
 import { toasts } from './toasts.svelte';
-import {
-	clamp,
-	compareBy,
-	entries,
-	mapValues,
-	pick,
-	safeJSONParse,
-	uint8ArrayToArrayBuffer
-} from './utils';
-
-/**
- * @param {DB.Observation[]} observations
- * @param {DB.Protocol} protocolUsed
- * @param {object} param2
- * @param {'croppedonly'|'full'|'metadataonly'} param2.include
- * @param {string} param2.base base path of the app - import `base` from `$app/paths`
- * @param {string} [param2.cropPadding] padding to add around the bounding box when cropping images. string of the form "npx" or "n%"
- */
-export async function generateResultsZip(
-	observations,
-	protocolUsed,
-	{ include = 'croppedonly', base, cropPadding }
-) {
-	const filepaths = protocolUsed.exports ?? {
-		images: {
-			cropped: Schemas.FilepathTemplate.assert('cropped/{{sequence}}.{{extension image.filename}}'),
-			original: Schemas.FilepathTemplate.assert(
-				'original/{{sequence}}.{{extension image.filename}}'
-			)
-		},
-		metadata: {
-			json: 'analysis.json',
-			csv: 'metadata.csv'
-		}
-	};
-
-	/**
-	 * @type {typeof Analysis.inferIn['observations']}
-	 */
-	let exportedObservations = {};
-	let sequence = 1;
-
-	// To have stable sequence numbers, really useful for testing
-	observations.sort(compareBy((o) => o.label + o.id));
-	for (const { id, label, images, metadataOverrides } of observations) {
-		const metadata = await observationMetadata({ images, metadataOverrides }).then(addValueLabels);
-
-		exportedObservations[id] = {
-			label,
-			metadata,
-			protocolMetadata: protocolMetadataValues(protocolUsed, metadata),
-			images: []
-		};
-
-		for (const imageId of images.sort()) {
-			const imageFromDatabase = db.tables.Image.state.find((i) => i.id === imageId);
-			if (!imageFromDatabase) continue;
-			const metadataValues = await addValueLabels(imageFromDatabase.metadata);
-
-			const image = {
-				...imageFromDatabase,
-				metadata: metadataValues,
-				protocolMetadata: protocolMetadataValues(protocolUsed, metadataValues)
-			};
-
-			const filepathsData = { observation: exportedObservations[id], image, sequence };
-
-			exportedObservations[id].images.push({
-				...image,
-				sequence,
-				exportedAs: {
-					original: filepaths.images.original.render(filepathsData),
-					cropped: filepaths.images.cropped.render(filepathsData)
-				}
-			});
-
-			sequence++;
-		}
-	}
-
-	const allMetadataKeys = [
-		// eslint-disable-next-line svelte/prefer-svelte-reactivity
-		...new Set(observations.flatMap((o) => Object.keys(exportedObservations[o.id].metadata)))
-	];
-
-	const metadataDefinitions = Object.fromEntries(db.tables.Metadata.state.map((m) => [m.id, m]));
-
-	/**
-	 * @type {Array<{imageId: string, croppedBytes: Uint8Array, originalBytes?: Uint8Array|undefined, contentType: string, filename: string}>}
-	 */
-	let buffersOfImages = [];
-
-	uiState.processing.state = 'generating-zip';
-	uiState.processing.total = 1;
-	uiState.processing.done = 0;
-
-	if (include !== 'metadataonly') {
-		uiState.processing.total += observations.flatMap((o) => o.images).length;
-		for (const [observation, imageId] of observations.flatMap((o) =>
-			o.images.map((img) => /** @type {const} */ ([o, img]))
-		)) {
-			const image = await db.tables.Image.get(imageId);
-			if (!image) continue;
-			const metadata = { ...image.metadata, ...observation.metadataOverrides };
-			const { contentType, filename } = image;
-			const { cropped, original } = await cropImage(protocolUsed, image, cropPadding);
-
-			/** @type {undefined | Uint8Array} */
-			let originalBytes = undefined;
-			/** @type {Uint8Array} */
-			let croppedBytes;
-
-			if (contentType === 'image/jpeg') {
-				croppedBytes = addExifMetadata(cropped, Object.values(metadataDefinitions), metadata);
-			} else {
-				croppedBytes = new Uint8Array(cropped);
-			}
-
-			if (include === 'full') {
-				if (contentType === 'image/jpeg') {
-					originalBytes = addExifMetadata(original, Object.values(metadataDefinitions), metadata);
-				} else {
-					originalBytes = new Uint8Array(original);
-				}
-			}
-
-			buffersOfImages.push({
-				imageId,
-				croppedBytes,
-				originalBytes,
-				contentType,
-				filename
-			});
-
-			uiState.processing.done++;
-		}
-	}
-
-	/**
-	 * @type {Uint8Array<ArrayBufferLike>}
-	 */
-	const zipfile = await new Promise((resolve, reject) =>
-		zip(
-			{
-				[filepaths.metadata.json]: strToU8(
-					stringifyWithToplevelOrdering(
-						'json',
-						`${window.location.origin}${base}/results.schema.json`,
-						Analysis.assert({
-							observations: exportedObservations,
-							protocol: {
-								...protocolUsed,
-								exports: {
-									...protocolUsed.exports,
-									images: {
-										original: filepaths.images.original.toJSON(),
-										cropped: filepaths.images.cropped.toJSON()
-									}
-								}
-							}
-						}),
-						['protocol', 'observations']
-					)
-				),
-				[filepaths.metadata.csv]: strToU8(
-					toCSV(
-						[
-							'Identifiant',
-							'Observation',
-							// 2 columns for each metadata: for the value itself, and for the confidence in the value
-							...allMetadataKeys
-								.filter((k) => Boolean(metadataDefinitions[k]?.label))
-								.flatMap((k) => [
-									metadataPrettyKey(metadataDefinitions[k]),
-									`${metadataPrettyKey(metadataDefinitions[k])}: Confiance`
-								])
-						],
-						observations.map((o) => ({
-							Identifiant: o.id,
-							Observation: o.label,
-							...Object.fromEntries(
-								Object.entries(exportedObservations[o.id].metadata).flatMap(
-									([key, { value, confidence, valueLabel }]) => [
-										[
-											metadataPrettyKey(metadataDefinitions[key]),
-											metadataPrettyValue(metadataDefinitions[key], value, valueLabel)
-										],
-										[
-											`${metadataPrettyKey(metadataDefinitions[key])}: Confiance`,
-											confidence.toString()
-										]
-									]
-								)
-							)
-						}))
-					)
-				),
-				...(include === 'metadataonly'
-					? {}
-					: Object.fromEntries(
-							Object.values(exportedObservations)
-								.flatMap(({ images }) => images)
-								.flatMap(({ exportedAs, id }) => {
-									const buffers = buffersOfImages.find((i) => i.imageId === id);
-									if (!buffers) return [];
-
-									return [
-										[exportedAs.cropped, [buffers.croppedBytes, { level: 0 }]],
-										[exportedAs.original, [buffers.originalBytes, { level: 0 }]]
-									].filter(([, [bytes]]) => bytes !== undefined);
-								})
-						))
-			},
-			{
-				// eslint-disable-next-line svelte/prefer-svelte-reactivity
-				comment: `Generated by C.i.g.a.l.e on ${new Date().toISOString()} - ${window.location.origin}`
-			},
-			(err, data) => {
-				if (err) reject(err);
-				resolve(data);
-			}
-		)
-	);
-
-	uiState.processing.done++;
-
-	downloadAsFile(zipfile, 'results.zip', 'application/zip');
-}
-
-/**
- * @param {DB.Protocol} protocol protocol used
- * @param {DB.Image} image
- * @param {string} [padding=protocol.crop.padding] padding to add around the bounding box when cropping images. string of the form "npx" or "n%"
-
- * @returns {Promise<{ cropped: ArrayBuffer, original: ArrayBuffer }>}
- */
-async function cropImage(protocol, image, padding = protocol.crop.padding) {
-	const centeredBoundingBox =
-		/** @type {undefined | import("./metadata").RuntimeValue<'boundingbox'>}  */
-		(image.metadata[protocol.crop?.metadata ?? 'crop']?.value);
-
-	const bytes = await db.get('ImageFile', imageIdToFileId(image.id)).then((f) => f?.bytes);
-	if (!bytes) throw "L'image n'a pas de fichier associé";
-
-	if (!centeredBoundingBox) {
-		return { cropped: bytes, original: bytes };
-	}
-
-	const bitmap = await createImageBitmap(new Blob([bytes], { type: image.contentType }));
-	const boundingBox = coordsScaler({ x: bitmap.width, y: bitmap.height })(
-		toTopLeftCoords(centeredBoundingBox)
-	);
-
-	const { inPixels } = parseCropPadding(padding);
-	const paddingPixels = { x: inPixels(bitmap.width), y: inPixels(bitmap.height) };
-
-	try {
-		const croppedBitmap = await createImageBitmap(
-			bitmap,
-			clamp(boundingBox.x - paddingPixels.x, 0, bitmap.width),
-			clamp(boundingBox.y - paddingPixels.y, 0, bitmap.height),
-			clamp(boundingBox.width + 2 * paddingPixels.x, 1, bitmap.width),
-			clamp(boundingBox.height + 2 * paddingPixels.y, 1, bitmap.height)
-		);
-		const canvas = document.createElement('canvas');
-		canvas.width = croppedBitmap.width;
-		canvas.height = croppedBitmap.height;
-		canvas.getContext('2d')?.drawImage(croppedBitmap, 0, 0);
-		const croppedBytes = await new Promise((resolve) =>
-			canvas.toBlob(
-				resolve,
-				['image/png', 'image/jpeg'].includes(image.contentType) ? image.contentType : 'image/png'
-			)
-		).then((blob) => blob.arrayBuffer());
-
-		// @ts-ignore
-		return { cropped: croppedBytes, original: bytes };
-	} catch (error) {
-		toasts.warn(`Impossible de recadrer ${image.filename}, l'image sera incluse sans recadrage`);
-		console.error(
-			`Couldn't crop ${image.filename} (id ${image.id}) with `,
-			{ boundingBox },
-			':',
-			error
-		);
-	} finally {
-		bitmap.close();
-	}
-
-	return { cropped: bytes, original: bytes };
-}
-
-/**
- *
- * @param {string} padding
- * @returns {{ withUnit: string, unitless: number, unit: 'px' | '%', inPixels: (axis: number) => number }}
- */
-export function parseCropPadding(padding) {
-	const match = padding.match(/(\d+)(px|%)/);
-	if (!match) throw new Error(`Invalid crop padding: ${padding}`);
-	const [value, unit] = [Number.parseInt(match[1], 10), /** @type {'%'|'px'} */ (match[2])];
-	return {
-		withUnit: padding,
-		unitless: value,
-		unit,
-		inPixels: (axis) => (unit === 'px' ? value : Math.round((axis * value) / 100))
-	};
-}
+import { entries, mapValues, pick, safeJSONParse, uint8ArrayToArrayBuffer } from './utils';
 
 /**
  * @template {string} HeaderKey
@@ -336,7 +14,7 @@ export function parseCropPadding(padding) {
  * @param {Array<Record<NoInfer<HeaderKey>, string>>} rows
  * @param {string} [separator=";"]
  */
-function toCSV(header, rows, separator = ';') {
+export function toCSV(header, rows, separator = ';') {
 	/** @param {string} cell */
 	const quote = (cell) => `"${cell?.replace(/"/g, '""') ?? ''}"`;
 
@@ -344,26 +22,6 @@ function toCSV(header, rows, separator = ';') {
 		header.map(quote).join(separator),
 		...rows.map((row) => header.map((key) => quote(row[key])).join(separator))
 	].join('\n');
-}
-
-/**
- * Returns a un-namespaced object of all metadata values of the given protocol, given the metadata values object of an image/observation. If a metadata value is absent from the given values, the value is still present, but set to `null`.
- *
- * @param {DB.Protocol} protocol
- * @param {DB.MetadataValues} values
- */
-function protocolMetadataValues(protocol, values) {
-	return Object.fromEntries(
-		protocol.metadata
-			.filter((key) => isNamespacedToProtocol(protocol.id, key))
-			.map((key) => [
-				removeNamespaceFromMetadataId(key),
-				values[key] ?? {
-					...METADATA_ZERO_VALUE,
-					valueLabel: ''
-				}
-			])
-	);
 }
 
 /**
